requirements.txt
flask==3.1.1
flask-session==0.8.0
cohere==5.15.0
python-dotenv==1.1.1
gunicorn==23.0.0